# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tickets API",
        "description": "API para gerenciamento de funcionários.",
        "contact": {
            "email": "jhonattan.curtarelli@outlook.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/employees": {
            "get": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Busca todos os funcionários",
                "operationId": "326f2e86fa40a9ba88b1cb5f2d4b62b8",
                "responses": {
                    "200": {
                        "description": "Lista de funcionários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Employee"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Cria um novo funcionário",
                "operationId": "4a10b3ada27736863f2586c56b9c19ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cpf"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "situation": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I"
                                        ],
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Funcionário cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erros de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dados Inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/employees/{id}": {
            "get": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Busca o funcionário pelo código",
                "operationId": "b7cb6100cedcac60d8c23d3f9d7c3251",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Código do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do funcionário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Funcionário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Funcionário não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Atualiza o funcionário",
                "operationId": "bcdffcad24dd066f489c0e6ed83861a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "situation": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I"
                                        ],
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Funcionário editado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erros de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dados Inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Deleta um funcionário",
                "operationId": "43aa0966dc246d0b0373a7db65a6bf78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Funcionário deletado com sucesso"
                    },
                    "400": {
                        "description": "Erro ao deletar o funcionário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao deletar o funcionário"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Employee": {
                "title": "Funcionários",
                "description": "Funcionários",
                "required": [
                    "id",
                    "name",
                    "cpf",
                    "situation"
                ],
                "properties": {
                    "id": {
                        "description": "ID do funcionário",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nome do funcionário",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "cpf": {
                        "description": "CPF do funcionário",
                        "type": "string",
                        "example": "12345678901"
                    },
                    "situation": {
                        "description": "Situação do funcionário",
                        "type": "string",
                        "enum": [
                            "A",
                            "I"
                        ],
                        "example": "A"
                    },
                    "created_at": {
                        "description": "Data de criação do funcionário",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-09-01T00:00:00"
                    },
                    "updated_at": {
                        "description": "Data de atualização do funcionário",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-09-01T00:00:00"
                    },
                    "deleted_at": {
                        "description": "Data de exclusão do funcionário",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-09-01T00:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}